Use cxxx headers in kernel when possible.

+++ kernel/include/arch/x86_64/arch.h
@@ -1,7 +1,7 @@
 #ifndef X86_64_ARCH_H
 #define X86_64_ARCH_H
 
-#include <sys/types.h>
+#include <cstdint>
 
 extern "C"
 {
@@ -9,7 +9,7 @@ extern "C"
 
     void flush_tlb(void);
 
-    void load_pml4(uint64_t addr);
+    void load_pml4(uintptr_t addr);
 }
 
 #endif // X86_64_ARCH_H
